Qn4 done: added put option binomial tree and american option

+++ computationalFinanceProject4/Mutils.h
@@ -14,7 +14,7 @@ public:
     static double Mean(long double arr[], int n);
     static double StDev(long double arr[], int n);
     static double Corr(long double *x, long double *y, int size);
-    static void WriteArrayToCSV(long double *arr, int n, const string filename);
+    static void WriteArrayToCSV( double *arr, int n, const string filename);
     static void WriteToCSV2DMatrix( double arr[][31], int col,int row, const string filename);
     static long double pnorm(double x);
     static double Cov(long double *x, long double *y, int size);
